sidecar: Do not overwrite files if contents haven't changed
If we update a label on the configmap/secret the sidecar will still
overwrite the file in the directory because it received an update event.
However, the contents haven't changed so we can avoid this operation.
This also has the additional benefit that other sidecars which happen to
watch the directory for changes will not trigger any reloads to the main
application.

The python implementation was inspired by [1] to ensure that we do not
chose on large reads

[1] https://www.quickprogrammingtips.com/python/how-to-calculate-sha256-hash-of-a-file-in-python.html