Add support for configmaps with binary data (#96)
k8s configmaps support binary data using the `binaryData` key. This is a
key/value pair just like the `data` key, but the contents are base64
encoded. To allow for compatibility with the changing of the function
definition, ASCII is set as the default.

Since the `binaryData` key is only in available in ConfigMaps the
blanket mapping of `sec.data` to `dataMap` was dropped and checks for
both data/binaryData had to be configured for only `ConfigMap` resources
while keeping `Secret` resources with the default check.

Iteration of keys and the data pass-through of value information to
`_get_file_data_and_name` also needed to be changed to use the resource
object directly instead of the `dataMap` variable.